[executor] Clean the structure of executor code
